FINDING PERFECT FEBRUARY
The length of the months and weeks do not always match each other. But there are years where February starts with Monday and ends with Sunday, which makes a perfect month for accountants.

For example, in 2021, 1st of February started with Monday, and lasted for 28 days, which is exactly 4 weeks.

We wonder if there are years between 1 and 2100 (including both of the years), which  contains perfect February month like the one described.

Write a python program to find the years that has a perfect February, which starts with Monday and has 28 days (i.e. not a leap year).

The program should output the years and as the last line, output the total number of years.